+++ src/wrappers.rs
@@ -61,6 +61,16 @@ impl<T: Identifiable> Boxed<T> {
     }
 }
 
+impl<T: Identifiable> Identifiable for Boxed<T> {
+    fn type_id(&self) -> u32 {
+        self.id
+    }
+
+    fn enum_variant_id(&self) -> Option<&'static str> {
+        None
+    }
+}
+
 impl<T: MtProtoSized> MtProtoSized for Boxed<T> {
     fn size_hint(&self) -> error::Result<usize> {
         let id_size_hint = self.id.size_hint()?;